Bug 457308 - Alternative mechanism to import projects
This alternative import framework allows user to start importing
project without knowing anythong of it. With this frameworks, Eclipse
projects for different technologies and languages register an extension
to detect and configure projects.
Assuming user has the right set of plugins installed in its IDE, result
is that they only have to provide as input a filesystem directory and
the framework will detect the various projects to import.
Basically this framework allows user to only care about WHAT to import
instead of caring about HOW and WHAT to import (current workflow).

Change-Id: I22d0d26786d0142dd0254fdfdbd188fc679fc156
Signed-off-by: Mickael Istria <mistria@redhat.com>