Bug 105372 - Builder to emulate classpath per source folder
This draft adds a builder that can be plugged on any Java project to
perform additional checks at build time. This builder uses additional
attributes in the .classpath definition to "restrict" the visibility of
classpath entries to specified source folders, and report an error if a
Java file inside a source folder reference a type that is from a
"restricted" classpath entry.

Change-Id: I9c69fad201c495f5c7b176024601e2463c44216b
Signed-off-by: Mickael Istria <mistria@redhat.com>